feat(scrape): unify layer scraping — one command owns layer/
`scrape layer` now scrapes both patterns and sessions via path-based
routing (LayerContent enum classifies layer/**/*.md files). Session
logic moved from scrape/sessions/ to scrape/layer/sessions.rs.

- `scrape layer` handles patterns + sessions in one pass
- `scrape sessions` shows deprecation warning, delegates to layer
- `scrape` default reduced from 5 steps to 4
- Beliefs excluded from router (separate embedding pipeline)
